Created subspec for bluetooth features
