+++ Sources/SwiftLocation/Extensions/Foundation+Extras.swift
@@ -24,7 +24,6 @@
 
 import Foundation
 import CoreLocation
-import CoreBluetooth
 
 // MARK: - Array
 
@@ -174,24 +173,6 @@ internal extension TimeInterval {
     
     static var highInterval: TimeInterval {
         3600
-    }
-    
-}
-
-// MARK: - CBManagerState
-
-extension CBManagerState: CustomStringConvertible {
-    
-    public var description: String {
-        switch self {
-        case .unknown: return "unknown"
-        case .resetting: return "resetting"
-        case .unsupported: return "unsupported"
-        case .unauthorized: return "unauthorized"
-        case .poweredOff: return "poweredOff"
-        case .poweredOn: return "poweredOn"
-        @unknown default: return ""
-        }
     }
     
 }
