+++ Sources/SwiftLocationBluetooth/Request/Requests/Beacon/BeaconBroadcaster.swift
