+++ Sources/SwiftLocationBluetooth/Request/Requests/Beacon/BeaconRequest.swift
