+++ SwiftLocation.podspec
@@ -1,6 +1,6 @@
 Pod::Spec.new do |s|
   s.name         = "SwiftLocation"
-  s.version      = "5.0.3"
+  s.version      = "5.1"
   s.summary      = "Location Manager Made Easy"
   s.description  = <<-DESC
   Efficient location tracking for iOS with support for oneshot/continuous/background tracking, reverse geocoding, autocomplete, geofencing, beacon monitoring & broadcasting
@@ -12,7 +12,12 @@ Pod::Spec.new do |s|
   s.ios.deployment_target = "11.0"
   s.macos.deployment_target = "11.0"
   s.source       = { :git => "https://github.com/malcommac/SwiftLocation.git", :tag => s.version.to_s }
-  s.source_files  = "Sources/**/*.swift"
+  s.source_files  = "Sources/SwiftLocation/**/*.swift"
   s.frameworks  = "Foundation","CoreLocation","MapKit"
   s.swift_versions = ['5.0', '5.1', '5.3']
+
+  s.subspec 'Main' do |s|
+    s.source_files = "Sources/SwiftLocationBluetooth/**/*.swift"
+    s.frameworks  = "Foundation","CoreBluetooth","MapKit"
+  end
 end
